% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/play_song.R
\name{error_songs}
\alias{error_songs}
\title{Rchestra}
\usage{
error_songs()
}
\description{
Plays a song from your web browser
}
\details{
This function specifically plays a song carefully curated for when you have hit a non-fatal error in R
}
\examples{
options(error = error_songs)
nonExistentVariableName
}
\seealso{
play_song
}

