% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/play_song.R
\name{play_song}
\alias{play_song}
\title{Rchestra}
\usage{
play_song(songName = "Never Gonna Give You Up Rick Astley", url = NULL,
  print = FALSE, site = "youtube.com")
}
\arguments{
\item{songName}{string used for google search. Typically including the song name and artist should be enough}

\item{url}{a string with a specific url to be launched. If left null, the function searches google instead}

\item{print}{boolean which tells whether to print the song being played to the console}

\item{site}{string which specifies which site to play from. By default it is www.youtube.com, but www.vevo.com (or any other
site of choice) can be used as well}
}
\description{
\code{play_song} Plays a song from your web browser
}
\details{
The function launches a new page in your web browser with the top hit from the google search for the
given string. By default, it selects the top result from www.youtube.com, but other options (such as vevo, etc)
can also be used
}
\examples{
play_song(songName = "Never Gonna Give You Up Rick Astley")
}
\seealso{
error_songs
}

