% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/play_song.R
\name{random_song}
\alias{random_song}
\title{Rchestra}
\usage{
random_song(playList = NULL, site = "youtube.com", seed = NULL)
}
\arguments{
\item{playList}{vector of strings containing phrases to search for on google}
}
\description{
Plays a song from your web browser
}
\details{
This function selects a random song from a specified playlist.
}
\examples{
random_song(c("Baby Justin Beiber", "Never Say Never Justin Beiber", "One Less Lonely Girl Justin Beiber"))
}
\seealso{
play_song
}

